Target difficulty updates
Several things have been changed/fixed in this commit:
- Target difficulty calculation now works for an arbitrary number of targets.
Including a calculation of target position on the unit sphere (see below),
difficulties for 1.2M targets can be calculated in under 1-2 mins.

- TaipanTargets now have a ucposn attribute for recording the position
of the target on the unit sphere. This is used for computing the difficulties
via a KDTree. This does not necessarily need to to be calculated at target
generation, as the compute_target_difficulties function will do that if and
when necessary.

- compute_target_difficulties has now been updated  to allow for the calculation
of a subset of difficulties, i.e. not all targets in the full list of targets
need updating. However, the full list of targets needs to be considered for the
difficulties to be correct. This extra parameter is optional, and will not affect
existing code.

- assign_tile, unpick_tile have been updated to correctly calculate the difficulty
of remaining targets, based on the update to compute_target_difficulties above

- _greedy tile generation has been fixed, so that targets within
TILE_RADIUS + 2*FIBRE_EXCLUSION_RADIUS of the selected tile centre have their
difficulties recomputed.