Don't let hours_obs for fields drop down to exactly 0
Allowing this would force that field to be observed, even if its score doesn't warrant it (and keep forcing that field to be observed whenever its up for the rest of the run). Instead, specify a minimum non-zero value allowed for hours_obs (i.e. the resolution of the almanacs being used). This at least gives the other components of the tile score a chance to have an effect.