Updates to priority logic based on ENT catalogue updates. Reverse change from 'visits' to 'observations' in most instances.
The reason that I went back to using 'visits' (i.e. visits this repeat) rather than 'observations' (total looks at the target) is to allow a target 'success' to be manually turned off, putting the target back in for full re-observation. In this instance, it is more appropriate to count the number of visits as the number of visits attempting to achieve 'success', rather than the total visits if we want a second full pass at the target. In the first instance (i.e. before 'success' is set to True by observations), 'visits'=='observations' in any case.